Refactor pages CRUD actions and create lists context menu.

+++ atlas-studio-service/src/main/java/com/kidscademy/atlas/studio/model/AtlasObject.java
@@ -259,9 +259,11 @@ public class AtlasObject implements RepositoryObject, HDateRange {
     @PreRemove
     public void preRemove() throws IOException {
 	File mediaDir = Files.objectDir(this);
-	Files.removeFilesHierarchy(mediaDir);
-	// Files.removeFilesHierarchy does not remove base directory itself
-	mediaDir.delete();
+	if (mediaDir.exists()) {
+	    Files.removeFilesHierarchy(mediaDir);
+	    // Files.removeFilesHierarchy does not remove base directory itself
+	    mediaDir.delete();
+	}
     }
 
     /**
